+++ CHANGELOG.md
@@ -42,6 +42,7 @@ dj.FreeTable(dj.conn(), "common_session.session_group").drop()
 - Remove numpy version restriction #1169
 - Merge table delete removes orphaned master entries #1164
 - Edit `merge_fetch` to expect positional before keyword arguments #1181
+- Allow part restriction `SpyglassMixinPart.delete` #1192
 
 ### Pipelines
 
@@ -53,9 +54,9 @@ dj.FreeTable(dj.conn(), "common_session.session_group").drop()
     - Export fixes #1164
         - Allow `get_abs_path` to add selection entry. #1164
         - Log restrictions and joins. #1164
-        - Check if querying table inherits mixin in `fetch_nwb`. #XXXX
-        - Ensure externals entries before adding to export. #XXXX
-    - Error specificity in `LabMemberInfo` #XXXX
+        - Check if querying table inherits mixin in `fetch_nwb`. #1192
+        - Ensure externals entries before adding to export. #1192
+    - Error specificity in `LabMemberInfo` #1192
 
 - Decoding